update metadata message
