+++ pkg/util/hostname/kubelet/host_alias.go
@@ -11,6 +11,7 @@ import (
 	"context"
 	"fmt"
 
+	"github.com/DataDog/datadog-agent/pkg/config"
 	"github.com/DataDog/datadog-agent/pkg/util/hostname/validate"
 )
 
@@ -21,4 +22,16 @@ func GetHostAlias(ctx context.Context) (string, error) {
 		return name, nil
 	}
 	return "", fmt.Errorf("couldn't extract a host alias from the kubelet: %s", err)
+}
+
+// GetMetaClusterNameText returns the clusterName text for the agent status output
+func GetMetaClusterNameText(ctx context.Context, hostname string) string {
+	if !config.IsFeaturePresent(config.Kubernetes) {
+		return ""
+	}
+	compliantClusterName, initialClusterName := getRFC1123CompliantClusterName(ctx, hostname)
+	if compliantClusterName != initialClusterName {
+		return fmt.Sprintf("%s (original name: %s)", compliantClusterName, initialClusterName)
+	}
+	return compliantClusterName
 }
